+++ subpackages/cli/source/odood/cli/commands/deploy.d
@@ -43,15 +43,23 @@ class CommandDeploy: OdoodCommand {
         this.add(new Option(
             null, "db-user", "Database port").defaultValue("odoo"));
         this.add(new Option(
-            null, "db-password", "Database password").defaultValue("odoo"));
+            null, "db-password", "Database password"));
         this.add(new Flag(
-            null, "local-postgres", "Configure local postgresql server"));
+            null, "local-postgres", "Configure local postgresql server (requires PostgreSQL installed)"));
 
         this.add(new Flag(
             null, "proxy-mode", "Enable proxy-mode in odoo config"));
 
         this.add(new Flag(
+            null, "local-nginx", "Autoconfigure local nginx (requires nginx installed)"));
+        this.add(new Option(
+            null, "local-nginx-server-name", "Servername for nginx config."));
+
+        this.add(new Flag(
             null, "enable-logrotate", "Enable logrotate for Odoo."));
+
+        this.add(new Flag(
+            null, "enable-fail2ban", "Enable fail2ban for Odoo (requires fail2ban installed)."));
 
         this.add(new Option(
             null, "supervisor", "What superwisor to use for deployment.")
@@ -101,6 +109,16 @@ class CommandDeploy: OdoodCommand {
 
         if (args.flag("enable-logrotate"))
             config.logrotate_enable = true;
+
+        if (args.flag("local-nginx")) {
+            config.nginx.enable = true;
+            config.nginx.server_name = args.option("local-nginx-server-name");
+            config.odoo.proxy_mode = true;
+            config.odoo.http_host = "127.0.0.1";
+        }
+
+        if (args.flag("enable-fail2ban"))
+            config.fail2ban_enable = true;
 
         if (args.option("supervisor"))
             switch(args.option("supervisor")) {